+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "google-auth-httplib2 (>=0.1.0,<1.0.0)",
     "google-auth-oauthlib (>=1.0.0,<2.0.0)",
     "cryptography (>=3.0.0,<4.0.0)",
+    "twilio (>=8.0.0,<9.0.0)",
 ]
